Move the utility function to get statuses into a PollbotAPI.js file
